+++ examples/debug.rs
@@ -8,8 +8,7 @@ fn main() {
     App::new()
         .add_plugins((
             DefaultPlugins.build().disable::<TransformPlugin>(),
-            BigSpacePlugin::default(),
-            FloatingOriginDebugPlugin::default(),
+            BigSpacePlugin::default().with_debug(),
         ))
         .add_systems(Startup, setup)
         .add_systems(Update, (movement, rotation))