+++ src/plugin.rs
@@ -1,24 +1,41 @@
 //! The bevy plugin for `big_space`.
 
-use crate::prelude::*;
-use alloc::vec::Vec;
+use crate::{prelude::*, timing::BigSpaceTimingStatsPlugin, validation::BigSpaceValidationPlugin};
 use bevy_app::prelude::*;
 use bevy_ecs::prelude::*;
 use bevy_transform::prelude::*;
 
-/// Add this plugin to your [`App`] for floating origin functionality.
+/// All plugins needed for the core functionality of [`big_space`](crate).
+///
+/// By default,
+/// - Hierarchy validation is enabled in debug, and disabled in release.
+/// - Debug gizmos are enabled if the feature is enabled.
+///
+/// This functionality can be disabled manually in this struct.
+///
+/// The debug plugin is behind a feature flag because, unlike hierarchy validation, it pulls in new
+/// dependencies.
 pub struct BigSpacePlugin {
     /// Enables runtime validation of `big_space` hierarchies, generating a detailed report on error
     /// using [`BigSpaceValidationPlugin`].
-    pub validate_hierarchies: bool,
-    /// When enabled, the crate runs all updates in the fixed timestep schedule.
-    pub fixed_timestep: bool,
+    pub validation: bool,
+    /// Enables debug gizmos to illustrate grid axes and occupied cells using
+    /// [`BigSpaceDebugPlugin`].
+    #[cfg(feature = "debug")]
+    pub debug: bool,
 }
 
 impl BigSpacePlugin {
-    /// Enable runtime hierarchy validation. See [`Self::validate_hierarchies`].
-    pub fn validate(mut self) -> Self {
-        self.validate_hierarchies = true;
+    /// Enable runtime hierarchy validation. See [`BigSpaceValidationPlugin`].
+    pub fn with_validation(mut self) -> Self {
+        self.validation = true;
+        self
+    }
+
+    /// Enable debug gizmos. See [`BigSpaceDebugPlugin`].
+    #[cfg(feature = "debug")]
+    pub fn with_debug(mut self) -> Self {
+        self.debug = true;
         self
     }
 }
@@ -26,8 +43,9 @@ impl BigSpacePlugin {
 impl Default for BigSpacePlugin {
     fn default() -> Self {
         Self {
-            validate_hierarchies: cfg!(debug_assertions),
-            fixed_timestep: false,
+            validation: cfg!(debug_assertions),
+            #[cfg(feature = "debug")]
+            debug: true,
         }
     }
 }
@@ -45,27 +63,22 @@ pub enum FloatingOriginSystem {
 impl Plugin for BigSpacePlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins((
-            BigSpaceMinimalPlugin {
-                fixed_timestep: self.fixed_timestep,
-            },
-            BigSpacePropagationPlugin {
-                fixed_timestep: self.fixed_timestep,
-            },
-            crate::timing::TimingStatsPlugin,
+            BigSpaceMinimalPlugin,
+            BigSpacePropagationPlugin,
+            BigSpaceTimingStatsPlugin,
         ));
-        if self.validate_hierarchies {
-            app.add_plugins(BigSpaceValidationPlugin {
-                fixed_timestep: self.fixed_timestep,
-            });
+        if self.validation {
+            app.add_plugins(BigSpaceValidationPlugin);
+        }
+        #[cfg(feature = "debug")]
+        if self.debug {
+            app.add_plugins(BigSpaceDebugPlugin);
         }
     }
 }
 
 /// Core setup needed for all uses of big space - reflection and grid cell recentering.
-pub struct BigSpaceMinimalPlugin {
-    /// When enabled, the plugin runs [`GridCell`] recentering in the fixed update schedule.
-    pub fixed_timestep: bool,
-}
+pub struct BigSpaceMinimalPlugin;
 impl Plugin for BigSpaceMinimalPlugin {
     fn build(&self, app: &mut App) {
         app.register_type::<Transform>()
@@ -74,36 +87,12 @@ impl Plugin for BigSpaceMinimalPlugin {
             .register_type::<GridCell>()
             .register_type::<Grid>()
             .register_type::<BigSpace>()
-            .register_type::<FloatingOrigin>();
-
-        let recenter_grid_cells = GridCell::recenter_large_transforms
-            .in_set(FloatingOriginSystem::RecenterLargeTransforms);
-
-        if self.fixed_timestep {
-            app.add_systems(FixedPostUpdate, recenter_grid_cells);
-        } else {
-            app.add_systems(PostUpdate, recenter_grid_cells);
-        }
-    }
-}
-
-/// Adds hierarchy validation features.
-pub struct BigSpaceValidationPlugin {
-    /// When enabled, the plugin runs validation in the fixed update schedule.
-    fixed_timestep: bool,
-}
-
-impl Plugin for BigSpaceValidationPlugin {
-    fn build(&self, app: &mut App) {
-        let config =
-            crate::validation::validate_hierarchy::<crate::validation::SpatialHierarchyRoot>
-                .after(TransformSystem::TransformPropagate);
-
-        if self.fixed_timestep {
-            app.add_systems(FixedPostUpdate, config);
-        } else {
-            app.add_systems(PostUpdate, config);
-        }
+            .register_type::<FloatingOrigin>()
+            .add_systems(
+                PostUpdate,
+                GridCell::recenter_large_transforms
+                    .in_set(FloatingOriginSystem::RecenterLargeTransforms),
+            );
     }
 }
 
@@ -113,11 +102,7 @@ impl Plugin for BigSpaceValidationPlugin {
 /// Disable Bevy's [`TransformPlugin`] while using this plugin.
 ///
 /// This also adds support for Bevy's low-precision [`Transform`] hierarchies.
-pub struct BigSpacePropagationPlugin {
-    /// When enabled, the plugin runs transform propagation in the fixed update schedule.
-    fixed_timestep: bool,
-}
-
+pub struct BigSpacePropagationPlugin;
 impl Plugin for BigSpacePropagationPlugin {
     fn build(&self, app: &mut App) {
         let configs = || {
@@ -140,13 +125,8 @@ impl Plugin for BigSpacePropagationPlugin {
                 .in_set(TransformSystem::TransformPropagate)
         };
 
-        app.add_systems(PostStartup, configs());
-
-        if self.fixed_timestep {
-            app.add_systems(FixedPostUpdate, configs());
-        } else {
-            app.add_systems(PostUpdate, configs());
-        };
+        app.add_systems(PostStartup, configs())
+            .add_systems(PostUpdate, configs());
 
         // These are the bevy transform propagation systems. Because these start from the root
         // of the hierarchy, and BigSpace bundles (at the root) do not contain a Transform,
@@ -160,7 +140,7 @@ impl Plugin for BigSpacePropagationPlugin {
         app.add_systems(
             PostStartup,
             (
-                propagate_parent_transforms,
+                crate::bevy_compat::propagate_parent_transforms,
                 bevy_transform::systems::sync_simple_transforms,
             )
                 .in_set(TransformSystem::TransformPropagate),
@@ -168,158 +148,10 @@ impl Plugin for BigSpacePropagationPlugin {
         .add_systems(
             PostUpdate,
             (
-                propagate_parent_transforms,
+                crate::bevy_compat::propagate_parent_transforms,
                 bevy_transform::systems::sync_simple_transforms,
             )
                 .in_set(TransformSystem::TransformPropagate),
         );
-    }
-}
-
-/// Copied from bevy. This is the simpler propagation implementation that doesn't use dirty tree
-/// marking. This is needed because dirty tree marking doesn't start from the root, and will end up
-/// doing the work for big space hierarchies, which it cannot affect anyway.
-pub fn propagate_parent_transforms(
-    mut root_query: Query<
-        (Entity, &Children, Ref<Transform>, &mut GlobalTransform),
-        Without<ChildOf>,
-    >,
-    mut orphaned: RemovedComponents<ChildOf>,
-    transform_query: Query<
-        (Ref<Transform>, &mut GlobalTransform, Option<&Children>),
-        With<ChildOf>,
-    >,
-    child_query: Query<(Entity, Ref<ChildOf>), With<GlobalTransform>>,
-    mut orphaned_entities: Local<Vec<Entity>>,
-) {
-    orphaned_entities.clear();
-    orphaned_entities.extend(orphaned.read());
-    orphaned_entities.sort_unstable();
-    root_query.par_iter_mut().for_each(
-        |(entity, children, transform, mut global_transform)| {
-            let changed = transform.is_changed() || global_transform.is_added() || orphaned_entities.binary_search(&entity).is_ok();
-            if changed {
-                *global_transform = GlobalTransform::from(*transform);
-            }
-
-            for (child, child_of) in child_query.iter_many(children) {
-                assert_eq!(
-                    child_of.parent, entity,
-                    "Malformed hierarchy. This probably means that your hierarchy has been improperly maintained, or contains a cycle"
-                );
-                // SAFETY:
-                // - `child` must have consistent parentage, or the above assertion would panic.
-                //   Since `child` is parented to a root entity, the entire hierarchy leading to it
-                //   is consistent.
-                // - We may operate as if all descendants are consistent, since
-                //   `propagate_recursive` will panic before continuing to propagate if it
-                //   encounters an entity with inconsistent parentage.
-                // - Since each root entity is unique and the hierarchy is consistent and
-                //   forest-like, other root entities' `propagate_recursive` calls will not conflict
-                //   with this one.
-                // - Since this is the only place where `transform_query` gets used, there will be
-                //   no conflicting fetches elsewhere.
-                #[expect(unsafe_code, reason = "`propagate_recursive()` is unsafe due to its use of `Query::get_unchecked()`.")]
-                unsafe {
-                    propagate_recursive(
-                        &global_transform,
-                        &transform_query,
-                        &child_query,
-                        child,
-                        changed || child_of.is_changed(),
-                    );
-                }
-            }
-        },
-    );
-}
-
-/// Recursively propagates the transforms for `entity` and all of its descendants.
-///
-/// # Panics
-///
-/// If `entity`'s descendants have a malformed hierarchy, this function will panic occur before
-/// propagating the transforms of any malformed entities and their descendants.
-///
-/// # Safety
-///
-/// - While this function is running, `transform_query` must not have any fetches for `entity`,
-///   nor any of its descendants.
-/// - The caller must ensure that the hierarchy leading to `entity` is well-formed and must
-///   remain as a tree or a forest. Each entity must have at most one parent.
-#[expect(
-    unsafe_code,
-    reason = "This function uses `Query::get_unchecked()`, which can result in multiple mutable references if the preconditions are not met."
-)]
-unsafe fn propagate_recursive(
-    parent: &GlobalTransform,
-    transform_query: &Query<
-        (Ref<Transform>, &mut GlobalTransform, Option<&Children>),
-        With<ChildOf>,
-    >,
-    child_query: &Query<(Entity, Ref<ChildOf>), With<GlobalTransform>>,
-    entity: Entity,
-    mut changed: bool,
-) {
-    let (global_matrix, children) = {
-        let Ok((transform, mut global_transform, children)) =
-            // SAFETY: This call cannot create aliased mutable references.
-            //   - The top level iteration parallelizes on the roots of the hierarchy.
-            //   - The caller ensures that each child has one and only one unique parent throughout
-            //     the entire hierarchy.
-            //
-            // For example, consider the following malformed hierarchy:
-            //
-            //     A
-            //   /   \
-            //  B     C
-            //   \   /
-            //     D
-            //
-            // D has two parents, B and C. If the propagation passes through C, but the ChildOf
-            // component on D points to B, the above check will panic as the origin parent does
-            // match the recorded parent.
-            //
-            // Also consider the following case, where A and B are roots:
-            //
-            //  A       B
-            //   \     /
-            //    C   D
-            //     \ /
-            //      E
-            //
-            // Even if these A and B start two separate tasks running in parallel, one of them will
-            // panic before attempting to mutably access E.
-            (unsafe { transform_query.get_unchecked(entity) }) else {
-            return;
-        };
-
-        changed |= transform.is_changed() || global_transform.is_added();
-        if changed {
-            *global_transform = parent.mul_transform(*transform);
-        }
-        (global_transform, children)
-    };
-
-    let Some(children) = children else { return };
-    for (child, child_of) in child_query.iter_many(children) {
-        assert_eq!(
-            child_of.parent, entity,
-            "Malformed hierarchy. This probably means that your hierarchy has been improperly maintained, or contains a cycle"
-        );
-        // SAFETY: The caller guarantees that `transform_query` will not be fetched for any
-        // descendants of `entity`, so it is safe to call `propagate_recursive` for each child.
-        //
-        // The above assertion ensures that each child has one and only one unique parent
-        // throughout the entire hierarchy.
-        unsafe {
-            propagate_recursive(
-                global_matrix.as_ref(),
-                transform_query,
-                child_query,
-                child,
-                changed || child_of.is_changed(),
-            );
-        }
     }
 }